mocr: fix deadlock when using a context across multiple threads
The first thread to initialize an mocr_ctx would never release the GIL.
When another thread would try to acquire the GIL this would create
deadlock.

This was fixed by adding a check for whether or not the GIL is held
after the call to Py_Initialize() in mocr_init(). If it was held, the
main interperter thread state is saved in a global variable so it can be
restored later if necessary. From there, nothing should block on
PyGILState_Ensure(), until it isn't safe.