mocr: add function to invoke finalize
Making this available from the API seems like the right thing to do
since you're supposed to free all the memory you allocate in a C
program. The reason I didn't do it though is because it gives the user
the ability to blow their whole leg off if they use it incorrectly.

With great power comes great responsibility. Don't use this method
unless you know what you are doing.